+++ app/src/main/AndroidManifest.xml
@@ -21,19 +21,16 @@
         android:theme="@style/Theme.HowTheyMove"
         tools:targetApi="31">
         <activity
+            android:name=".EnergyListElement"
+            android:exported="false" />
+        <activity
+            android:name=".EnergyList"
+            android:exported="false" />
+        <activity
             android:name=".MoreInformations"
             android:exported="false" />
         <activity
-            android:name=".act2"
-            android:exported="false" />
-        <activity
-            android:name=".act1"
-            android:exported="false" />
-        <activity
             android:name=".Custom_List"
-            android:exported="false" />
-        <activity
-            android:name=".Definitii"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
