+++ app/src/main/java/com/example/howtheymove/EnergyListElement.java
@@ -4,11 +4,11 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class Custom_List extends AppCompatActivity {
+public class EnergyListElement extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_custom_list);
+        setContentView(R.layout.activity_energy_list_element);
     }
 }
