--- app/src/main/java/com/example/howtheymove/act1.java
@@ -1,14 +1,0 @@
-package com.example.howtheymove;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class act1 extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_act1);
-    }
-}
