+++ app/src/main/res/layout/activity_energy_list_element.xml
